@@ -63,28 +64,28 @@
     "msw": "latest",
     "msw-storybook-addon": "canary",
     "prettier": "^3.2.5",
-    "prisma": "^5.10.2",
-    "projen": "^0.80.7",
+    "prisma": "^5.11.0",
+    "projen": "^0.80.10",
     "semantic-release-plus": "^20.0.0",
-    "storybook": "^8.0.0-alpha.13",
+    "storybook": "^8.0.0",
     "storybook-addon-module-mock": "^1.1.5",
     "storybook-addon-turbo-build": "^2.0.1",
     "ts-node": "^10.9.2",
     "tsconfig-paths-webpack-plugin": "^4.1.0",
-    "typescript": "^5.3.3"
+    "typescript": "^5.4.2"
   },
   "dependencies": {
-    "@aws-sdk/client-s3": "^3.525.0",
-    "@aws-sdk/s3-request-presigner": "^3.525.0",
+    "@aws-sdk/client-s3": "^3.533.0",
+    "@aws-sdk/s3-request-presigner": "^3.533.0",
     "@chakra-ui/icons": "^2.1.1",
     "@chakra-ui/next-js": "^2.2.0",
     "@chakra-ui/react": "^2.8.2",
     "@chakra-ui/theme-tools": "^2.1.2",
     "@egjs/react-flicking": "^4.11.2",
-    "@elastic/react-search-ui": "^1.20.2",
-    "@elastic/react-search-ui-views": "^1.20.2",
-    "@elastic/search-ui": "^1.20.2",
-    "@elastic/search-ui-elasticsearch-connector": "^1.20.2",
+    "@elastic/react-search-ui": "^1.21.0",
+    "@elastic/react-search-ui-views": "^1.21.0",
+    "@elastic/search-ui": "^1.21.0",
+    "@elastic/search-ui-elasticsearch-connector": "^1.21.0",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.0",
     "@hookform/resolvers": "^3.3.4",
