@@ -3,22 +3,6 @@ import { ImSpinner8 } from "react-icons/im";
 import { cx, css } from "~/styled-system/css";
 import { button, type ButtonVariantProps } from "~/styled-system/recipes";
 
-function LoaderIcon() {
-  return (
-    <span
-      className={css({
-        position: "absolute",
-        top: "50%",
-        left: "50%",
-        transform: "translate(-50%, -50%)",
-      })}
-      aria-hidden
-    >
-      <ImSpinner8 className={css({ animation: "spin" })} />
-    </span>
-  );
-}
-
 export const buttonColorPalette = [
   "primary",
   "secondary",
@@ -26,14 +10,13 @@ export const buttonColorPalette = [
   "success",
 ] as const;
 
-export type ButtonProps = {
-  icon?: React.ReactNode;
-  isLoading?: boolean;
-  iconPosition?: "left" | "right";
-  colorPalette?: (typeof buttonColorPalette)[number];
-  isIconButton?: boolean;
-} & ButtonVariantProps &
-  ComponentProps<"button">;
+export type ButtonProps = ButtonVariantProps &
+  ComponentProps<"button"> & {
+    icon?: React.ReactNode;
+    isLoading?: boolean;
+    iconPosition?: "left" | "right";
+    colorPalette?: (typeof buttonColorPalette)[number];
+  };
 
 export function Button({
   icon = null,
@@ -55,15 +38,39 @@ export function Button({
       disabled={isLoading || disabled}
       {...restProps}
     >
-      {!isLoading && icon && <span aria-hidden>{icon}</span>}
+      {icon && (
+        <span
+          className={css({ _loading: { opacity: 0 } })}
+          aria-busy={isLoading}
+          aria-hidden
+        >
+          {icon}
+        </span>
+      )}
       <span
-        className={css({ "&[data-loading=true]": { opacity: 0 } })}
-        data-loading={isLoading}
+        className={css({ _loading: { opacity: 0 } })}
+        aria-busy={isLoading}
         data-label
       >
         {children}
       </span>
       {isLoading && <LoaderIcon />}
     </button>
+  );
+}
+
+function LoaderIcon() {
+  return (
+    <span
+      className={css({
+        position: "absolute",
+        top: "50%",
+        left: "50%",
+        transform: "translate(-50%, -50%)",
+      })}
+      aria-hidden
+    >
+      <ImSpinner8 className={css({ animation: "spin" })} />
+    </span>
   );
 }
